requirements.txt
celery==5.2.7
redis==4.3.4
Django==4.1.1
django-bootstrap-icons==0.8.1
django-crispy-forms==1.14.0
Pillow==9.2.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
uWSGI==2.0.20
